Split MQTT from main library
Commit 2de38d1 ("mqtt") added MQTT support but did so in a way that
makes the basic Texecom connect functionality impossible to use without
MQTT. Split the generic functionality to a separate library. Change
alarm-monitor.py to derive a new class from the base which overloads
necessary functions to add mqtt functionality back.
There should be no functional change introduced by this commit.
